refactor: widgets/streaming 컴포넌트 ui 폴더로 이동
