+++ client/src/widgets/streaming/ui/AlbumBackground.tsx
