+++ client/src/widgets/streaming/ui/LyricsPanel.css
