+++ client/src/widgets/streaming/ui/Streaming.tsx
